tty: n_smux: Enable IPC logging for SMUX
Using the kernel log for SMUX logging requires users to enable
debug logging and then reproduce the issue.  Enabling all
logging often affects the timing and can make issues harder
to solve.

Add SMUX IPC Logging to enable internal logging that has a
minimum impact on performance and is always on to allow
debugging of an issue either from the command line or from
a memory dump.

CRs-Fixed: 384568
Change-Id: I6f1df407e07760206a7a548c9d3fce228e249155
Signed-off-by: Angshuman Sarkar <angshuman@codeaurora.org>

Conflicts:
	drivers/tty/n_smux.c