coresight: change to using smp call for etm enable/disable
Since we moved to separate enable nodes for each ETM, using pm_qos
to prevent power collapse of the target cpu while we enable or
disable ETM is less efficient and also susceptible to a race with
cpu hotplug.

Move to using smp call to enable and disable trace to avoid above
deficiencies with the newly re-structured code.

Change-Id: I65c726dbd6275af8db0d40db5034299287a02e91
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>

Conflicts:
	drivers/coresight/coresight-etm.c