media: dvb: mpq: Secure demux support
This change adds support for demuxing of encrypted streams which current
software demux implementation does not support.
User-space API is extended to allow setting of key ladder information
needed in order to perform decryption.
If secure demux application is available, demuxing does not take place in
the dvb demux sw filter anymore but instead is performed in the new secure
demux, regardless of whether the stream is encrypted or not.
Otherwise, the dvb demux sw filter will be used as before and demuxing of
encrypted streams will not be available.

Change-Id: If75a7b03d1f13d3b012250b87c25eaf6d6cbc399
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>

Conflicts:
	drivers/media/dvb/dvb-core/demux.h
	drivers/media/dvb/dvb-core/dmxdev.c
	drivers/media/dvb/dvb-core/dvb_demux.c
	drivers/media/dvb/dvb-core/dvb_demux.h
	drivers/media/dvb/mpq/adapter/mpq_stream_buffer.c
	drivers/media/dvb/mpq/demux/mpq_dmx_plugin_common.c
	drivers/media/dvb/mpq/demux/mpq_dmx_plugin_common.h
	drivers/media/dvb/mpq/demux/mpq_dmx_plugin_tspp_v1.c
	include/linux/dvb/dmx.h