media: dvb: mpq: Fix issue with framing mechanism
The demux driver identifies video frames by searchning for patterns in
the video payload. A pattern can be spread over two separate TS packets,
so the pattern lookup mechanism must take into consideration the possible
prefix sizes. Also, when reporting the written data length, the prefix
size (if used) should be added.
This patch fixes issues with not taking all possible prefix sizes into
consideration, and with not reporting the data length properly when a
prefix is used.

Change-Id: I6f4829e860c2d7518964a456a36418be26a06080
Signed-off-by: Liron Kuch <lkuch@codeaurora.org>

Conflicts:
	drivers/media/dvb/mpq/demux/mpq_dmx_plugin_common.c