msm: mdm: do not assert ap2mdm_errfatal gpio in some cases during SSR
Some platforms have the ap2mdm_errfatal line shared between
two external modems. Current implementation is to assert that
gpio during the SSR shutdown callback but that does not allow
for independent restart of each modem. On most platforms it is
actually not necessary to assert the ap2mdm_errfatal during
restart of the mdm. It is necessary during kernel panic of the
apps processor but that is already done in the panic handler.
The use case where it is necessary to assert the errfatal gpio
is where the mdm needs to be restarted by the SSR module, for
example on platforms where the internal and external modems need
to be restarted together. This patch adds a platform data parameter
to identify those platforms.

Change-Id: Ie4f5cbed752a04a80f1e4bfb50da03f52e010045
Signed-off-by: Joel King <joelking@codeaurora.org>