msm: camera: Fix the composite stats buffer handling.
Currently we configure composite stats in HFR mode. In composite stats
mode, the stats buffers are not getting dequeued and dispatched properly.
Hence the stats are not processed properly in HFR mode, resulting in dark
preview in camcorder HFR mode. Fix this by correcting the buffer handling
logic in case of composite stats.

CRs-Fixed: 452197
Change-Id: I98be1160001deea72d0fa31e0e804d523fd72c20
Signed-off-by: Ashwini Rao <ashwinik@codeaurora.org>
Signed-off-by: Kiran Kumar H N <hurlisal@codeaurora.org>