tspp: Add synchronization between API call and tasklet
The TSPP driver uses a tasklet which is scheduled on BAM transfer
completion to process the data. The tasklet verifies that the
channel had not already been closed. Therefore, it is necessary
to have a synchronization mechanism between the channel_close API
function and the tasklet function. Added acquisition of spinlock
in the channel_close function.

Change-Id: I274298cbc9c83b4b6c1d50ffb709b8c5d8d43231
CRs-Fixed: 477556
Signed-off-by: Liron Kuch <lkuch@codeaurora.org>

Conflicts:
	drivers/misc/tspp.c